LPS-175383 Move code to the new component and update accordingly
