+++ modules/apps/fragment/fragment-renderer-react-impl/src/main/java/com/liferay/fragment/renderer/react/internal/model/listener/FragmentEntryLinkModelListener.java
@@ -43,60 +43,10 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Iván Zaera Avellón
  */
-@Component(
-	service = {
-		FragmentEntryLinkModelListener.class, IdentifiableOSGiService.class,
-		ModelListener.class
-	}
-)
+@Component(service = {IdentifiableOSGiService.class, ModelListener.class})
 public class FragmentEntryLinkModelListener
 	extends BaseModelListener<FragmentEntryLink>
 	implements IdentifiableOSGiService {
-
-	public void ensureInitialized() {
-		if (_initialized) {
-			return;
-		}
-
-		synchronized (this) {
-			if (_initialized) {
-				return;
-			}
-
-			JSPackage jsPackage = _npmResolver.getJSPackage();
-
-			if (jsPackage == null) {
-				if (_log.isDebugEnabled()) {
-					_log.debug(
-						"Unable to initialize because JS package is null");
-				}
-
-				return;
-			}
-
-			List<FragmentEntryLink> fragmentEntryLinks =
-				_fragmentEntryLinkLocalService.getFragmentEntryLinks(
-					FragmentConstants.TYPE_REACT, QueryUtil.ALL_POS,
-					QueryUtil.ALL_POS, null);
-
-			NPMRegistryUpdate npmRegistryUpdate = _npmRegistry.update();
-
-			for (FragmentEntryLink fragmentEntryLink : fragmentEntryLinks) {
-				npmRegistryUpdate.registerJSModule(
-					jsPackage,
-					FragmentEntryFragmentRendererReactUtil.getModuleName(
-						fragmentEntryLink),
-					FragmentEntryFragmentRendererReactUtil.getDependencies(),
-					FragmentEntryFragmentRendererReactUtil.getJs(
-						fragmentEntryLink, jsPackage),
-					null);
-			}
-
-			npmRegistryUpdate.finish();
-
-			_initialized = true;
-		}
-	}
 
 	@Override
 	public String getOSGiServiceIdentifier() {
@@ -246,8 +196,6 @@ public class FragmentEntryLinkModelListener
 
 	@Reference
 	private FragmentEntryLinkLocalService _fragmentEntryLinkLocalService;
-
-	private volatile boolean _initialized;
 
 	@Reference
 	private NPMRegistry _npmRegistry;
