+++ modules/apps/fragment/fragment-renderer-react-impl/src/main/java/com/liferay/fragment/renderer/react/internal/renderer/FragmentEntryFragmentRendererReact.java
@@ -20,7 +20,7 @@ import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.fragment.renderer.FragmentRendererContext;
 import com.liferay.fragment.renderer.constants.FragmentRendererConstants;
-import com.liferay.fragment.renderer.react.internal.model.listener.FragmentEntryLinkModelListener;
+import com.liferay.fragment.renderer.react.internal.helper.FragmentEntryLinkJSModuleInitializerHelper;
 import com.liferay.fragment.renderer.react.internal.util.FragmentEntryFragmentRendererReactUtil;
 import com.liferay.fragment.util.configuration.FragmentEntryConfigurationParser;
 import com.liferay.frontend.js.loader.modules.extender.npm.JSPackage;
@@ -96,7 +96,7 @@ public class FragmentEntryFragmentRendererReact implements FragmentRenderer {
 			HttpServletResponse httpServletResponse)
 		throws IOException {
 
-		_fragmentEntryLinkModelListener.ensureInitialized();
+		_fragmentEntryLinkJSModuleInitializerHelper.ensureInitialized();
 
 		try {
 			PrintWriter printWriter = httpServletResponse.getWriter();
@@ -134,6 +134,8 @@ public class FragmentEntryFragmentRendererReact implements FragmentRenderer {
 	@Activate
 	protected void activate() {
 		_jsPackage = _npmResolver.getJSPackage();
+
+		_fragmentEntryLinkJSModuleInitializerHelper.ensureInitialized();
 	}
 
 	private FragmentEntry _getContributedFragmentEntry(
@@ -257,7 +259,8 @@ public class FragmentEntryFragmentRendererReact implements FragmentRenderer {
 	private FragmentEntryConfigurationParser _fragmentEntryConfigurationParser;
 
 	@Reference
-	private FragmentEntryLinkModelListener _fragmentEntryLinkModelListener;
+	private FragmentEntryLinkJSModuleInitializerHelper
+		_fragmentEntryLinkJSModuleInitializerHelper;
 
 	@Reference
 	private JSONFactory _jsonFactory;
